+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/intake/Intake.java
@@ -34,7 +34,8 @@ public class Intake extends SubsystemBase {
      */
     public Intake(HardwareMap hardwareMap) {
         intakeMotor = hardwareMap.get(DcMotorEx.class, IntakeConstants.intakeMotorName);
-        // Direction: FORWARD (default)
+        // Direction: REVERSE (motor is mounted inverted)
+        intakeMotor.setDirection(DcMotorSimple.Direction.REVERSE);
         isRunning = true; // Intake runs continuously by default after initialization
     }
 
@@ -167,15 +168,15 @@ public class Intake extends SubsystemBase {
 
             // Priority: fastShooting > fullPower > standard
             if (isShooting) {
-                targetPower = IntakeConstants.transitPower; //1
+                targetPower = - IntakeConstants.transitPower; //1
             } else if (isFastIntaking){
-                targetPower = IntakeConstants.fastIntakePower; // 0.7
+                targetPower = - IntakeConstants.fastIntakePower; // 0.7
             } else if (isFastShooting) {
-                targetPower = IntakeConstants.fastShootingPower; // 0.8
+                targetPower = - IntakeConstants.fastShootingPower; // 0.8
             } else if (isFullPower) {
                 targetPower = IntakeConstants.fullPower; // 0.65
             } else {
-                targetPower = IntakeConstants.intakePower; // 0.5
+                targetPower = - IntakeConstants.intakePower; // 0.5
             }
 
             // Reverse logic